refactor: 모든 출력은 OutputManger 가 하도록 변경
