+++ src/main/java/racingcar/io/InputManager.java
@@ -1,24 +1,25 @@
 package racingcar.io;
 
-import static racingcar.io.ErrorMessage.INVALID_TRY_COUNT_MESSAGE;
-import static racingcar.io.ErrorMessage.INVALID_USER_INPUT_MESSAGE;
-import static racingcar.io.ErrorMessage.NOT_POSITIVE_NUMBER_ERROR_MESSAGE;
+import static racingcar.io.constant.ErrorMessage.ATTEMPT_COUNT_MESSAGE;
+import static racingcar.io.constant.ErrorMessage.GUIDE_MESSAGE;
+import static racingcar.io.constant.ErrorMessage.INVALID_TRY_COUNT_MESSAGE;
+import static racingcar.io.constant.ErrorMessage.INVALID_USER_INPUT_MESSAGE;
+import static racingcar.io.constant.ErrorMessage.NOT_POSITIVE_NUMBER_ERROR_MESSAGE;
 
 import camp.nextstep.edu.missionutils.Console;
 
 public class InputManager {
 
     public static String readUserInput() {
-        OutputManager.printGuideMessage();
+        OutputManager.printMessage(GUIDE_MESSAGE);
+
         String input = Console.readLine();
-        if (input.isBlank()) {
-            throw new IllegalArgumentException(INVALID_USER_INPUT_MESSAGE);
-        }
+        validateEmptyString(input);
         return input;
     }
 
     public static int readAttemptCount() {
-        OutputManager.printTryCountMessage();
+        OutputManager.printMessage(ATTEMPT_COUNT_MESSAGE);
 
         int attemptCount = 0;
 
@@ -31,6 +32,12 @@ public class InputManager {
 
         validatePositiveNumber(attemptCount);
         return attemptCount;
+    }
+
+    private static void validateEmptyString(String input) {
+        if (input.isBlank()) {
+            throw new IllegalArgumentException(INVALID_USER_INPUT_MESSAGE);
+        }
     }
 
     private static void validatePositiveNumber(int attemptCount) {
