+++ src/main/java/racingcar/io/OutputManager.java
@@ -1,8 +1,7 @@
 package racingcar.io;
 
-import static racingcar.io.ErrorMessage.GUIDE_MESSAGE;
-import static racingcar.io.ErrorMessage.ATTEMPT_COUNT_MESSAGE;
-import static racingcar.io.ErrorMessage.WINNER_RESULT_MESSAGE;
+import static racingcar.io.constant.ErrorMessage.ATTEMPT_COUNT_MESSAGE;
+import static racingcar.io.constant.ErrorMessage.WINNER_RESULT_MESSAGE;
 
 import java.util.List;
 
@@ -11,10 +10,6 @@ public class OutputManager {
 
     private static final String DELIMITER = ", ";
 
-    public static void printGuideMessage() {
-        System.out.println(GUIDE_MESSAGE);
-    }
-
     public static void printWinner(List<String> winnerList) {
         String joinedName = String.join(DELIMITER, winnerList);
         System.out.println(WINNER_RESULT_MESSAGE + joinedName);
@@ -22,5 +17,13 @@ public class OutputManager {
 
     public static void printTryCountMessage() {
         System.out.println(ATTEMPT_COUNT_MESSAGE);
+    }
+
+    public static void printMessage(String message) {
+        System.out.println(message);
+    }
+
+    public static void printMessage() {
+        System.out.println();
     }
 }
